requirements.txt
litellm>=1.55.0
python-dotenv>=1.0.0
garminconnect>=0.2.22
requests>=2.32.0
websocket-client>=1.8.0
